+++ src/bigbang/dispatcher.h
@@ -18,9 +18,9 @@ public:
     ~CDispatcher();
     Errno AddNewBlock(const CBlock& block, uint64 nNonce = 0) override;
     Errno AddNewTx(const CTransaction& tx, uint64 nNonce = 0) override;
-    bool AddNewDistribute(const uint256& hashAnchor, const CDestination& dest,
+    bool AddNewDistribute(const int& hashAnchor, const CDestination& dest,
                           const std::vector<unsigned char>& vchDistribute) override;
-    bool AddNewPublish(const uint256& hashAnchor, const CDestination& dest,
+    bool AddNewPublish(const int& hashAnchor, const CDestination& dest,
                        const std::vector<unsigned char>& vchPublish) override;
 
 protected:
