+++ src/bigbang/struct.h
@@ -148,7 +148,7 @@ public:
 class CDelegateRoutine
 {
 public:
-    std::vector<std::pair<uint256, std::map<CDestination, size_t>>> vEnrolledWeight;
+    std::vector<std::pair<int, std::map<CDestination, size_t>>> vEnrolledWeight;
 
     std::vector<CTransaction> vEnrollTx;
     std::map<CDestination, std::vector<unsigned char>> mapDistributeData;
